refactor : refactor parseTimeStamp done
